@@ -7,22 +7,22 @@
     "@emotion/cache": "^11.11.0",
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
-    "@mui/icons-material": "^5.2.5",
-    "@mui/material": "^5.2.6",
+    "@mui/icons-material": "^5.16.4",
+    "@mui/material": "^5.16.4",
     "@rollup/plugin-typescript": "^11.1.6",
     "@types/d3-drag": "^3.0.7",
     "colord": "^2.9.3",
     "copy-anything": "^3.0.3",
-    "d3": "^7.8.4",
+    "d3": "^7.9.0",
     "d3-drag": "^3.0.0",
     "msw": "^0.35.0",
     "nanoid": "^3.3.4",
-    "node-fetch": "^2.6.9",
+    "node-fetch": "^3.3.2",
     "pluralize": "^8.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "recoil": "^0.7.7",
-    "rollup-plugin-dotenv": "^0.4.1",
+    "rollup-plugin-dotenv": "^0.5.1",
     "sleep-promise": "^9.1.0",
     "togostanza": "github:togostanza/togostanza",
     "yohak-tools": "0.1.18"
@@ -41,7 +41,7 @@
   },
   "private": true,
   "devDependencies": {
-    "@babel/core": "^7.24.7",
+    "@babel/core": "^7.24.9",
     "@emotion/babel-plugin": "^11.11.0",
     "@emotion/babel-preset-css-prop": "^11.11.0",
     "@emotion/eslint-plugin": "^11.11.0",
@@ -53,8 +53,8 @@
     "@types/d3": "^7.4.0",
     "@types/jest": "^27.4.0",
     "@types/pluralize": "0.0.29",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
     "@typescript-eslint/eslint-plugin": "^7.13.1",
     "@typescript-eslint/parser": "^7.13.1",
     "babel-loader": "^9.1.3",
@@ -69,8 +69,8 @@
     "jest": "^29.7.0",
     "jest-extended": "^4.0.2",
     "msw-storybook-addon": "^1.5.0",
-    "prettier": "^3.3.2",
-    "react-docgen-typescript-plugin": "^1.0.5",
+    "prettier": "^3.3.3",
+    "react-docgen-typescript-plugin": "^1.0.8",
     "storybook-addon-paddings": "^4.2.1",
     "ts-jest": "^29.1.5",
     "tsc-watch": "^4.2.9",
