@@ -3,7 +3,7 @@ import dotenv from "rollup-plugin-dotenv";
 export default function config(environment) {
   return {
     rollup: {
-      plugins: [dotenv.default()],
+      plugins: [dotenv()],
     },
   };
 }
